+++ src/internal.ts
@@ -177,4 +177,4 @@ export { revokeIdentityToCreatePortfolios } from '~/api/procedures/revokeIdentit
 export { rotatePrimaryKeyToSecondary } from '~/api/procedures/rotatePrimaryKeyToSecondary';
 export { registerCustomAssetType } from '~/api/procedures/registerCustomAssetType';
 export { bondPolyx } from '~/api/procedures/bondPolyx';
-export { unbondPolyx } from '~/api/procedures/unbondPolyx';
+export { updateBondedPolyx } from '~/api/procedures/updateBondedPolyx';
